+++ app/controllers/users_controller.rb
@@ -4,19 +4,18 @@ class UsersController < ApplicationController
     end
 
     def create
-        puts params
         @user = User.new(params.require(:user).permit(:username,:email,:password))
         if @user.save
             flash[:notice] = "Welcome #{@user.username}"
             session[:user_id] = @user.id
-            redirect_to user_path(@user)
+            redirect_to articles_path
         else
             render 'new'
         end
     end
 
     def edit
-        
+        #ensures a user is editing his own account by checking the id
         if logged_in? && current_user.id.to_s == params[:id]
             @user = User.find(params[:id])
         else 
@@ -26,12 +25,17 @@ class UsersController < ApplicationController
     end
     
     def update
-        @user = User.find(params[:id])
-        if @user.update(params.require(:user).permit(:username,:email,:password))
-            flash[:notice] = "Updated successfully"
-            redirect_to articles_path
+        if logged_in? && current_user.id.to_s == params[:id]
+            @user = User.find(params[:id])
+            if @user.update(params.require(:user).permit(:username,:email,:password))
+                flash[:notice] = "Updated successfully"
+                redirect_to articles_path
+            else
+                render 'edit'
+            end
         else
-            render 'edit'
+            flash[:alert] = "You can only update your account"
+            redirect_to root_path
         end
     end
 
@@ -45,13 +49,16 @@ class UsersController < ApplicationController
 
     def destroy
         #delete all his articles,delete the user and logout him
-        user = current_user
-        puts "###################################################################################"
-        puts user
-        puts "###################################################################################"
-        user.destroy
-        session[:user_id] = nil
-        flash[:notice] = "Account deleted"
-        redirect_to login_path
+        if logged_in? && current_user.id.to_s == params[:id]
+            user = current_user
+            user.destroy
+            session[:user_id] = nil
+            flash[:notice] = "Account deleted"
+            redirect_to login_path
+        else
+            flash[:alert] = "You can only delete your account"
+            redirect_to root_path
+        end
+
     end
 end
